+++ pykeg/src/pykeg/connections/twitter/urls.py
@@ -1,0 +1,6 @@
+from django.conf import settings
+from django.conf.urls.defaults import *
+
+urlpatterns = patterns('pykeg.connections.twitter.views',
+)
+