+++ pykeg/src/pykeg/web/context_processors.py
@@ -3,12 +3,7 @@ from pykeg.core import models
 from pykeg.web.kegweb.forms import LoginForm
 
 def enabled_features(request):
-  """Adds a USE_FEATURENAME flags for each enabled feature.
-
-  Currently defined:
-    USE_FACEBOOK - Allow registration and fbconnect login via django-socialauth
-    USE_TWITTER - pykeg.contrib.twitter is installed.
-  """
+  """Adds a USE_FEATURENAME flags for each enabled feature (see features.py)"""
   # TODO(mikey): this might make it harder to diagnose why features aren't
   # visible/being used.
   ret = {}